Warning problems (Thanks to Jan Springer for the patch!)

+++ Source/System/NodeCores/Drawables/VolRen/testMtexLUTShaderRender.cpp
@@ -195,12 +195,12 @@ void display(void)
     {
 	// Eventually switch between different hardware modes
 	float ctime   = glutGet(GLUT_ELAPSED_TIME);
-	int   dtime   = ctime / 1000 / 2;
+	int   dtime   = (int)ctime / 1000 / 2;
 
 	newMode = dtime % numModes;
 
 	// Make the volume rotate
- 	int movement = (int) (ctime - lastFrame) * aniSpeed;
+ 	int movement = (int) ((ctime - lastFrame) * aniSpeed);
  	if (movement != 0) lastFrame = ctime;
  
 	mgr->mouseButtonPress  (GLUT_LEFT_BUTTON,  0, 0);
